working core data

+++ ios_new/TalkieWalkie/Stores/User.swift
@@ -7,16 +7,20 @@
 
 import CoreData
 import Foundation
+import OSLog
 
 import FirebaseAuth
 
 extension User {
     @discardableResult
-    static func upsert(_ u: App_User, context: NSManagedObjectContext) -> User {
-        let localU = User(context: context)
+    static func fromProto(_ u: App_User, context: NSManagedObjectContext, block: (_ user: User) -> Void = { _ in }) -> User {
+        let localU = User.getByUuidOrCreate(u.uuid.uuidOrThrow(), context: context)
         localU.uuid = u.uuid.uuidOrThrow()
         localU.displayName = u.displayName
         localU.phone = u.phone
+    
+        block(localU)
+        
         return localU
     }
 }
@@ -27,13 +31,16 @@ extension Me {
         req.predicate = .all
 
         let res = (try? context.fetch(req)) ?? []
-        
-        // I hate Core Data. Why would we have multiple instance of the Me object???
-        context.perform {
-            res.filter { $0.uuid == nil }.forEach { context.delete($0) }
-            context.saveOrLogError()
+
+        if res.count > 1 {
+            os_log(.error, "getting \(res.count) instances of [Me] object in core data.")
+            // I hate Core Data. Why would we have multiple instance of the Me object???
+            context.perform {
+                res.filter { $0.uuid == nil }.forEach { context.delete($0) }
+                context.saveOrLogError()
+            }
         }
-        
+
         return res.first { $0.uuid != nil }
     }
 }
